+++ NeoForge/src/main/java/fuzs/puzzleslib/neoforge/impl/data/NeoForgeTagAppenderV2.java
@@ -12,16 +12,10 @@ import org.jetbrains.annotations.Nullable;
 import java.util.List;
 import java.util.function.Function;
 
-public final class NeoForgeTagAppender<T> extends AbstractTagAppender<T> {
+public final class NeoForgeTagAppenderV2<T> extends AbstractTagAppender<T> {
 
-    public NeoForgeTagAppender(TagBuilder tagBuilder, @Nullable Function<T, ResourceKey<T>> keyExtractor) {
+    public NeoForgeTagAppenderV2(TagBuilder tagBuilder, @Nullable Function<T, ResourceKey<T>> keyExtractor) {
         super(tagBuilder, keyExtractor);
-    }
-
-    @Override
-    public AbstractTagAppender<T> setReplace(boolean replace) {
-        this.tagBuilder.replace(replace);
-        return this;
     }
 
     @Override
@@ -40,12 +34,10 @@ public final class NeoForgeTagAppender<T> extends AbstractTagAppender<T> {
     public List<String> asStringList() {
         ImmutableList.Builder<String> builder = ImmutableList.builder();
         for (TagEntry tagEntry : this.tagBuilder.build()) {
-            builder.add(new ExtraCodecs.TagOrElementLocation(
-                    tagEntry.getId(), tagEntry.isTag()).toString());
+            builder.add(new ExtraCodecs.TagOrElementLocation(tagEntry.getId(), tagEntry.isTag()).toString());
         }
         for (TagEntry tagEntry : this.tagBuilder.getRemoveEntries().toList()) {
-            builder.add("!" + new ExtraCodecs.TagOrElementLocation(
-                    tagEntry.getId(), tagEntry.isTag()));
+            builder.add("!" + new ExtraCodecs.TagOrElementLocation(tagEntry.getId(), tagEntry.isTag()));
         }
         return builder.build();
     }